--- src/App.jsx
@@ -1,7 +1,0 @@
-import SignUpPage from './page/SignUpPage';
-
-function App() {
-  return <SignUpPage />;
-}
-
-export default App;
